+++ x-pack/solutions/observability/plugins/slo/server/routes/utils/get_slo_server_route_repository.ts
@@ -10,7 +10,7 @@ import type { batchGetCompositeSLORoute } from '../composite_slo/batch_get_compo
 import type { createCompositeSLORoute } from '../composite_slo/create_composite_slo';
 import type { getCompositeSLORoute } from '../composite_slo/get_composite_slo';
 import type { getCompositeSLOSuggestionsRoute } from '../composite_slo/get_composite_slo_suggestions';
-import type { findCompositeSLORoute } from '../composite_slo/find_composite_slo';
+import type { findCompositeSLORoute } from '../composite_slo/find_composite_slo_definitions';
 import type { updateCompositeSLORoute } from '../composite_slo/update_composite_slo';
 import type { deleteCompositeSLORoute } from '../composite_slo/delete_composite_slo';
 import type { fetchCompositeHistoricalSummaryRoute } from '../composite_slo/fetch_composite_historical_summary';
